+++ src/routes/headerMenuItems.js
@@ -1,0 +1,43 @@
+const headerMenuItems = [
+  {
+    title: "關於我們",
+    link: "/about",
+    subItems: [
+      { title: "師資", link: "/about#teachers" },
+      { title: "交通資訊", link: "/about#transport" },
+    ],
+  },
+  {
+    title: "探索課程",
+    link: "/explore-courses",
+    subItems: [
+      { title: "運動保健", link: "/explore-courses#sport" },
+      { title: "心靈成長", link: "/explore-courses#mind" },
+      { title: "生活技能", link: "/explore-courses#skills" },
+      { title: "社區活動", link: "/explore-courses#community" },
+    ],
+  },
+  {
+    title: "知識分享",
+    link: "/knowledge-sharing",
+    subItems: [
+      { title: "運動", link: "/knowledge-sharing#sport" },
+      { title: "飲食", link: "/knowledge-sharing#food" },
+      { title: "養生", link: "/knowledge-sharing#health" },
+      { title: "醫療", link: "/knowledge-sharing#medical" },
+      { title: "癌症", link: "/knowledge-sharing#cancer" },
+      { title: "居家", link: "/knowledge-sharing#home" },
+    ],
+  },
+  {
+    title: "活動照片",
+    link: "/photos",
+    subItems: [
+      { title: "講座", link: "/photos#lecture" },
+      { title: "社團", link: "/photos#club" },
+      { title: "運動會", link: "/photos#sportsevent" },
+    ],
+  },
+];
+
+export default headerMenuItems;
